requirements.txt
# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0  # For neural network zone classification
torch>=2.0.0
torchaudio>=2.0.0
pandas>=2.0.0

# Metadata Handling
mutagen>=1.47.0

# GUI Framework
PyQt5>=5.15.0
pyqtgraph>=0.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Performance
numba>=0.57.0  # JIT compilation for librosa

# Telegram Bot
aiogram>=3.0.0
python-dotenv>=1.0.0

# Task Queue (Redis 8.x/Valkey compatible)
arq>=0.25.0
redis>=5.0.0
